fix: Resolve overflow and gap issues on mobile for getting started page download boxes
Previously the getting started page download boxes overflowed their container and had a weird gap on mobile devices. These changes contain the cards inside the same container width as the rest of the page and give an even gap around each side of each card on desktop and mobile devices.