chore: starting to work out 'Dtga-Title-text'
Created DtgA-Title-Text early in the project, this is my first ever html/css/javascript project and I've learned that I don't like that particular approach. Much easier and more understandable to just put the color in the individual classes in the css.

I'll keep going through and changing this from having the color from that tag.